Move CommandNotFound to exceptions.py
