+++ dredis/lua.py
@@ -2,7 +2,8 @@ import json
 
 from lupa._lupa import LuaRuntime
 
-from dredis.commands import run_command, CommandNotFound, SimpleString
+from dredis.commands import run_command, SimpleString
+from dredis.exceptions import CommandNotFound
 
 
 class RedisScriptError(Exception):
